Support for GetDynamicAsync
The other GetX methods are structured. That is, they accept only supported query parameters and return structured results. This is because the query parameters affect the resulting output (e.g, keys & raw).

If you want to use a single method call you can use the GetDynamicAsync method which accepts any query parameters (for example, using anonymous types to pass any query parameters). The trade-off is the result is dynamic as well and the caller has to deal with the varying content.